Kraken: refactoring for add test on "use_crow_fly"
Add some forward declaration in raptor API, so I have to add header in a
lot of files
