+++ source/routing/raptor_api.h
@@ -29,14 +29,24 @@ www.navitia.io
 */
 
 #pragma once
-#include "raptor.h"
 #include "type/type.pb.h"
 #include "type/response.pb.h"
 #include "type/request.pb.h"
-#include "boost/date_time/posix_time/ptime.hpp"
-#include "georef/street_network.h"
+#include <limits>
+
+namespace navitia{
+    namespace type{
+        class EntryPoint;
+        class AccessibiliteParams;
+    }
+    namespace georef{
+        class StreetNetwork;
+    }
+}
 
 namespace navitia { namespace routing {
+
+class RAPTOR;
 
 pbnavitia::Response make_response(RAPTOR &raptor,
                                   const type::EntryPoint &origin,
