+++ source/routing/routing.h
@@ -117,6 +117,14 @@ struct Path {
 
 bool operator==(const PathItem & a, const PathItem & b);
 
+/**
+ * Choose if we must use a crowfly or a streetnework for a section. This function is call for the first and last section of a journey
+ *
+ * @param point: the object where we are going/coming: the requested origin for the first section or the destination for the last section
+ * @param stop_point: for the first section, the stop point where we are going, or for the last section the stop point from where we come
+ */
+bool use_crow_fly(const type::EntryPoint& point, const type::StopPoint* stop_point);
+
 
 }}
